test(17-06): wire fs/s3/memory backends to blockstoretest + delete legacy testkits
- fs + local-memory backends call BlockStoreConformance + BlockStoreAppendConformance
- s3 + remote-memory backends call BlockStoreConformance
- pkg/blockstore/local/localtest/ DELETED (951 LoC)
- pkg/blockstore/remote/remotetest/ DELETED (644 LoC)
- backend-specific fs scenarios preserved next to backend
