+++ pkg/blockstore/local/fs/fs_conformance_test.go
@@ -4,36 +4,72 @@ import (
 	"context"
 	"testing"
 
+	"github.com/marmos91/dittofs/pkg/blockstore"
+	"github.com/marmos91/dittofs/pkg/blockstore/blockstoretest"
 	"github.com/marmos91/dittofs/pkg/blockstore/local/fs"
-	"github.com/marmos91/dittofs/pkg/blockstore/local/localtest"
 	memmeta "github.com/marmos91/dittofs/pkg/metadata/store/memory"
 )
 
-// TestFSStore_AppendLogConformance runs the shared D-22 append-log
-// conformance suite (LSL-01..LSL-06 + INV-03/INV-05) against the
-// in-tree *fs.FSStore. The factory constructs a fresh store with
-// UseAppendLog=true, a memory-backed RollupStore, and the rollup pool
-// started — matching the production wiring contract Phase 11 (A2) will
-// flip the default on.
-func TestFSStore_AppendLogConformance(t *testing.T) {
-	localtest.RunAppendLogSuite(t, func(t *testing.T) *fs.FSStore {
-		t.Helper()
-		dir := t.TempDir()
-		rs := memmeta.NewMemoryMetadataStoreWithDefaults()
-		bc, err := fs.NewWithOptions(dir, 1<<30, 1<<30, nil, fs.FSStoreOptions{
-			UseAppendLog:    true,
-			MaxLogBytes:     1 << 30,
-			RollupWorkers:   2,
-			StabilizationMS: 50,
-			RollupStore:     rs,
-		})
-		if err != nil {
-			t.Fatalf("NewWithOptions: %v", err)
-		}
-		if err := bc.StartRollup(context.Background()); err != nil {
-			t.Fatalf("StartRollup: %v", err)
-		}
-		t.Cleanup(func() { _ = bc.Close() })
-		return bc
+// newFSStoreForConformance builds a fresh *fs.FSStore with the production
+// wiring used across Phase 17 conformance scenarios: UseAppendLog=true,
+// memory-backed RollupStore, rollup pool started. Returns the store and a
+// cleanup closure that closes it. Plan 17-06 D-09: shared between the
+// BlockStore and BlockStoreAppend factories so the wiring is identical
+// across both contracts.
+func newFSStoreForConformance(t *testing.T) *fs.FSStore {
+	t.Helper()
+	dir := t.TempDir()
+	rs := memmeta.NewMemoryMetadataStoreWithDefaults()
+	bc, err := fs.NewWithOptions(dir, 1<<30, 1<<30, nil, fs.FSStoreOptions{
+		UseAppendLog:    true,
+		MaxLogBytes:     1 << 30,
+		RollupWorkers:   2,
+		StabilizationMS: 50,
+		RollupStore:     rs,
 	})
+	if err != nil {
+		t.Fatalf("NewWithOptions: %v", err)
+	}
+	if err := bc.StartRollup(context.Background()); err != nil {
+		t.Fatalf("StartRollup: %v", err)
+	}
+	return bc
+}
+
+// TestFSStore_BlockStoreConformance runs the unified Phase 17 D-09
+// BlockStoreConformance suite from pkg/blockstore/blockstoretest against
+// *fs.FSStore. The fs backend satisfies the BlockStore contract by way
+// of its embedded BlockStoreAppend implementation (Plan 17-07 lands the
+// missing Put/Has/Walk/etc. methods that complete the interface).
+func TestFSStore_BlockStoreConformance(t *testing.T) {
+	factory := func(t *testing.T) (blockstore.BlockStore, func()) {
+		t.Helper()
+		bc := newFSStoreForConformance(t)
+		cleanup := func() { _ = bc.Close() }
+		return bc, cleanup
+	}
+	blockstoretest.BlockStoreConformance(t, factory)
+}
+
+// TestFSStore_BlockStoreAppendConformance runs the random-write
+// absorber suite from pkg/blockstore/blockstoretest against
+// *fs.FSStore. The fs backend implements BlockStoreAppend in full (it
+// is the only backend that does so per D-09 — s3 / memory-remote
+// implement only BlockStore).
+//
+// Note: three of the five scenarios (PressureChannel_INV05,
+// TornWriteRecovery_LSL06, RollupOffsetMonotone_INV03) `t.Skip` on the
+// interface-only surface because they require fs-internal probes that
+// intentionally do not appear on BlockStoreAppend (Plan 17-02 SUMMARY
+// §"Three legacy scenarios `t.Skip` with explicit rationale"). The fs
+// backend continues to exercise those scenarios in
+// appendlog_test.go / recovery_test.go via internal test hooks.
+func TestFSStore_BlockStoreAppendConformance(t *testing.T) {
+	factory := func(t *testing.T) (blockstore.BlockStoreAppend, func()) {
+		t.Helper()
+		bc := newFSStoreForConformance(t)
+		cleanup := func() { _ = bc.Close() }
+		return bc, cleanup
+	}
+	blockstoretest.BlockStoreAppendConformance(t, factory)
 }
